Little Forest 상세도감 수정

+++ MyLittleForRest/app/src/main/java/com/beehivestudio/mylittleforrest/PopupActivity.java
@@ -6,9 +6,17 @@ import android.os.Bundle;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.Window;
+import android.widget.TextView;
+
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 public class PopupActivity extends Activity {
 
+    private String menu;
+    private String contentUid;
+    private String selectOption;
+    FirebaseFirestore db = FirebaseFirestore.getInstance();
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -16,18 +24,40 @@ public class PopupActivity extends Activity {
         requestWindowFeature(Window.FEATURE_NO_TITLE);
         setContentView(R.layout.activity_popup);
 
+        TextView head = findViewById(R.id.head);
+        TextView txtText = findViewById(R.id.txtText);
+        menu = getIntent().getStringExtra("menu");
+        selectOption = getIntent().getStringExtra("selectOption");
+
+        contentUid = getIntent().getStringExtra("contentUid");
+        if(menu.equals("delete")){
+            head.setText("삭제");
+            txtText.setText("정말 지우시겠습니까?");
+        }
+
     }
 
     public void yes(View v){
-        Intent intent = new Intent(this, SeedActivity.class);
-        startActivity(intent);
-        finish();
+        if (menu.equals("delete")){
+            Intent intent =new Intent(this, ContentDetailActivity.class);
+            db.collection(selectOption).document(contentUid).delete();
+            setResult(RESULT_OK, intent);
+            finish();
+        }else {
+            Intent intent = new Intent(this, SeedActivity.class);
+            startActivity(intent);
+            finish();
+        }
     }
 
     public void no(View v){
-        moveTaskToBack(true);
-        finish();
-        android.os.Process.killProcess(android.os.Process.myPid());
+        if (menu.equals("delete")){
+            finish();
+        }else {
+            moveTaskToBack(true);
+            finish();
+            android.os.Process.killProcess(android.os.Process.myPid());
+        }
     }
 
     @Override
